small_examples/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
mlflow==3.2.0
databricks-vectorsearch==0.57
databricks-langchain==0.5.1
langchain-core==0.3.74